Refactor split large test
